modules/user.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-
from telegram import \
    ReplyKeyboardMarkup,\
    ReplyKeyboardRemove,\
    Update

from telegram.ext import \
    Updater,\
    CommandHandler,\
    MessageHandler,\
    Filters,\
    ConversationHandler,\
    CallbackContext

UserConversationHandler = ConversationHandler(
    entry_points=[],
    states={},
    fallbacks=[],
    per_chat=True,
)
