requirements.txt
coloredlogs
python-telegram-bot
pymongo
